backend/pyproject.toml
[project]
name = "portfolio-backend"
version = "0.1.0"
description = "Portfolio backend API"
authors = [
    { name = "Luc Joosten", email = "lhajoosten@outlook.com" }
]
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.37",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pgvector>=0.3.6",
    # Validation & settings
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "email-validator>=2.2.0",
    # Auth
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.20",
    # AI
    "openai>=1.60.0",
    # HTTP
    "httpx>=0.28.0",
]

[dependency-groups]
dev = [
    "ruff>=0.9.0",
    "mypy>=1.14.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",  # for TestClient
    "testcontainers[postgres]>=4.8.0",
    "pytest-docker>=3.1.0",
]
docs = [
    # MkDocs core + Material theme
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
    # Auto-generate API reference from Python docstrings
    "mkdocstrings[python]>=0.25.0",
    # Programmatic page generation
    "mkdocs-gen-files>=0.5.0",
    # Auto-build nav from generated SUMMARY.md
    "mkdocs-literate-nav>=0.6.0",
    # Allow index pages for nav sections
    "mkdocs-section-index>=0.3.9",
    # Mermaid diagram support
    "mkdocs-mermaid2-plugin>=1.1.1",
    # Git revision / last-updated dates
    "mkdocs-git-revision-date-localized-plugin>=1.2.0",
    # Minify HTML/JS/CSS output
    "mkdocs-minify-plugin>=0.8.0",
    # PyMdown Extensions (required by Material theme features)
    "pymdown-extensions>=10.7.1",
    # Syntax highlighting
    "pygments>=2.17.0",
]

[project.optional-dependencies]
email = [
    "email-validator>=2.2.0",
]

# =============================================================================
# Ruff — linting + formatting (replaces black, isort, flake8)
# =============================================================================
[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "RUF", # ruff-specific
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults (FastAPI Depends)
    "N999",   # invalid module name
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# =============================================================================
# MyPy — type checking
# =============================================================================
[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true

# =============================================================================
# Pytest
# =============================================================================
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
